Fix Tcs calculation in v10 and v11 miits analysis jupyter notebooks
Current sharing temperature should be calculated considering only
the area of the wire which is superconducting. The stabilizer portion,
which in our case is copper, should not be considered for current
density calculation when calculating the current sharing temperature
for the given wire.

This commit fix this for v10 and v11 MIITS analysis.
This change decreases the composite thermal capacity
and conductivity of the wire (considering copper plus Nb-Ti),
but increases the quench propagation velocity.
Overall, for the tested cases, this change did not increase the
hot-spot temperature, since the decreased capacity was compensated
by earlier quench detection.